chore(geo): split out testData and testFunctions
